pyproject.toml
[project]
name = "intellicredit"
version = "0.1.0-mvp"
description = "IntelliCredit MVP - Automated Credit Analysis"
readme = "README_MVP.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "streamlit>=1.25.0",
    "loguru>=0.7.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "pdfplumber>=0.10.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "altair>=5.0.0",
    "openai>=1.40.0",
    "anthropic>=0.31.0",
    "httpx>=0.27",
    "tenacity>=8.2",
    "tiktoken>=0.7",
    "markdownify>=0.12",
    "orjson>=3.10",
    "jsonschema>=4.22",
    "python-slugify>=8",
    "databricks-sdk>=0.28",
    "pyarrow>=15",
    "sqlalchemy>=2.0",
    "python-docx>=1.0.1",
    "reportlab>=4.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
